+++ src/models/product-variant-model.js
@@ -1,5 +1,5 @@
 import BaseModel from './base-model';
-
+import ImageModel from './image-model';
 
 /**
   * Model for product variant
@@ -150,7 +150,13 @@ const ProductVariantModel = BaseModel.extend({
       return image.variant_ids.indexOf(id) !== -1;
     })[0];
 
-    return (variantImage || primaryImage);
+    const productImage = variantImage || primaryImage;
+
+    if (!productImage) {
+      return null;
+    }
+
+    return new ImageModel(productImage);
   },
 
   /**
@@ -174,34 +180,11 @@ const ProductVariantModel = BaseModel.extend({
     * @type {Array}
   */
   get imageVariants() {
-    const image = this.image;
-
-    if (!image) {
+    if (!this.image) {
       return [];
     }
 
-    const src = this.image.src;
-    const extensionIndex = src.lastIndexOf('.');
-    const pathAndBasename = src.slice(0, extensionIndex);
-    const extension = src.slice(extensionIndex);
-    const variants = [
-      { name: 'pico', dimension: '16x16' },
-      { name: 'icon', dimension: '32x32' },
-      { name: 'thumb', dimension: '50x50' },
-      { name: 'small', dimension: '100x100' },
-      { name: 'compact', dimension: '160x160' },
-      { name: 'medium', dimension: '240x240' },
-      { name: 'large', dimension: '480x480' },
-      { name: 'grande', dimension: '600x600' },
-      { name: '1024x1024', dimension: '1024x1024' },
-      { name: '2048x2048', dimension: '2048x2048' }
-    ];
-
-    variants.forEach(variant => {
-      variant.src = `${pathAndBasename}_${variant.name}${extension}`;
-    });
-
-    return variants;
+    return this.image.variants;
   },
 
   /**
